Update logistic_regression.Rmd
Added a simple matrix illustration of the linear model (thought it fitted well with the text immediately before it).

Changed the name of 'appetite_d' to 'appetite_dummy' - just for readability (less for the reader to hold in mind?).

Added a brief nod to the fact that the odds ratio can be calculated with counts or proportions. (Sometimes proportions are mentioned in the text, sometimes counts, so thought it would be good to mention it explicitly).

Added an illustration (in a dataframe) of how the odds ratio changes as the number of events/nonevents changes.

Added a scatter plot showing how the likelihood is high for good predictions are low for bad predictions.